# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto List Bahrain API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CarFuel = (typeof CarFuel)[keyof typeof CarFuel];

export const CarFuel = {
  petrol: "petrol",
  diesel: "diesel",
  hybrid: "hybrid",
  electric: "electric",
} as const;

export type CarTransmission =
  (typeof CarTransmission)[keyof typeof CarTransmission];

export const CarTransmission = {
  automatic: "automatic",
  manual: "manual",
} as const;

export type CarBodyType = (typeof CarBodyType)[keyof typeof CarBodyType];

export const CarBodyType = {
  sedan: "sedan",
  suv: "suv",
  coupe: "coupe",
  hatchback: "hatchback",
  convertible: "convertible",
  pickup: "pickup",
  van: "van",
  wagon: "wagon",
} as const;

export type CarCondition = (typeof CarCondition)[keyof typeof CarCondition];

export const CarCondition = {
  new: "new",
  used: "used",
} as const;

export interface Car {
  id: number;
  make: string;
  model: string;
  trim?: string | null;
  year: number;
  priceBhd: number;
  mileageKm: number;
  fuel: CarFuel;
  transmission: CarTransmission;
  bodyType: CarBodyType;
  color: string;
  condition: CarCondition;
  location: string;
  sellerName: string;
  sellerPhone: string;
  description: string;
  images: string[];
  featured: boolean;
  createdAt: string;
}

export type CarInputFuel = (typeof CarInputFuel)[keyof typeof CarInputFuel];

export const CarInputFuel = {
  petrol: "petrol",
  diesel: "diesel",
  hybrid: "hybrid",
  electric: "electric",
} as const;

export type CarInputTransmission =
  (typeof CarInputTransmission)[keyof typeof CarInputTransmission];

export const CarInputTransmission = {
  automatic: "automatic",
  manual: "manual",
} as const;

export type CarInputBodyType =
  (typeof CarInputBodyType)[keyof typeof CarInputBodyType];

export const CarInputBodyType = {
  sedan: "sedan",
  suv: "suv",
  coupe: "coupe",
  hatchback: "hatchback",
  convertible: "convertible",
  pickup: "pickup",
  van: "van",
  wagon: "wagon",
} as const;

export type CarInputCondition =
  (typeof CarInputCondition)[keyof typeof CarInputCondition];

export const CarInputCondition = {
  new: "new",
  used: "used",
} as const;

export interface CarInput {
  make: string;
  model: string;
  trim?: string | null;
  year: number;
  priceBhd: number;
  mileageKm: number;
  fuel: CarInputFuel;
  transmission: CarInputTransmission;
  bodyType: CarInputBodyType;
  color: string;
  condition: CarInputCondition;
  location: string;
  sellerName: string;
  sellerPhone: string;
  description: string;
  images: string[];
}

export interface MarketplaceSummary {
  totalListings: number;
  averagePriceBhd: number;
  newCount: number;
  usedCount: number;
  makes: number;
}

export interface MakeCount {
  make: string;
  count: number;
}

export interface BodyTypeCount {
  bodyType: string;
  count: number;
}

export interface Inquiry {
  id: number;
  carId: number;
  name: string;
  phone: string;
  email?: string | null;
  message: string;
  createdAt: string;
}

export interface InquiryInput {
  carId: number;
  name: string;
  phone: string;
  email?: string | null;
  message: string;
}

export type ListCarsParams = {
  q?: string;
  make?: string;
  bodyType?: string;
  fuel?: string;
  transmission?: string;
  condition?: ListCarsCondition;
  minPrice?: number;
  maxPrice?: number;
  minYear?: number;
  maxYear?: number;
  sort?: ListCarsSort;
  limit?: number;
  offset?: number;
};

export type ListCarsCondition =
  (typeof ListCarsCondition)[keyof typeof ListCarsCondition];

export const ListCarsCondition = {
  new: "new",
  used: "used",
} as const;

export type ListCarsSort = (typeof ListCarsSort)[keyof typeof ListCarsSort];

export const ListCarsSort = {
  newest: "newest",
  price_asc: "price_asc",
  price_desc: "price_desc",
  year_desc: "year_desc",
  mileage_asc: "mileage_asc",
} as const;

export type ListCars200 = {
  items: Car[];
  total: number;
};

export type GetRecentCarsParams = {
  limit?: number;
};