# lib/api-zod/src/generated/types/car.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto List Bahrain API
 * OpenAPI spec version: 0.1.0
 */
import type { CarBodyType } from "./carBodyType";
import type { CarCondition } from "./carCondition";
import type { CarFuel } from "./carFuel";
import type { CarTransmission } from "./carTransmission";

export interface Car {
  id: number;
  make: string;
  model: string;
  trim?: string | null;
  year: number;
  priceBhd: number;
  mileageKm: number;
  fuel: CarFuel;
  transmission: CarTransmission;
  bodyType: CarBodyType;
  color: string;
  condition: CarCondition;
  location: string;
  sellerName: string;
  sellerPhone: string;
  description: string;
  images: string[];
  featured: boolean;
  createdAt: Date;
}