# lib/api-zod/src/generated/types/carInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto List Bahrain API
 * OpenAPI spec version: 0.1.0
 */
import type { CarInputBodyType } from "./carInputBodyType";
import type { CarInputCondition } from "./carInputCondition";
import type { CarInputFuel } from "./carInputFuel";
import type { CarInputTransmission } from "./carInputTransmission";

export interface CarInput {
  make: string;
  model: string;
  trim?: string | null;
  year: number;
  priceBhd: number;
  mileageKm: number;
  fuel: CarInputFuel;
  transmission: CarInputTransmission;
  bodyType: CarInputBodyType;
  color: string;
  condition: CarInputCondition;
  location: string;
  sellerName: string;
  sellerPhone: string;
  description: string;
  images: string[];
}