# lib/api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto List Bahrain API
 * OpenAPI spec version: 0.1.0
 */

export interface InquiryInput {
  carId: number;
  name: string;
  phone: string;
  email?: string | null;
  message: string;
}