# lib/api-zod/src/generated/types/marketplaceSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto List Bahrain API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplaceSummary {
  totalListings: number;
  averagePriceBhd: number;
  newCount: number;
  usedCount: number;
  makes: number;
}